BME160 Final Project - Relevant Research on FastA
Intro:
Todd Lowe’s GtRNAdb (genomic tRNA database) is one of the world’s leading research collections for tRNA’s. A vast number of researchers and scientists use the GtRNAdb as a reliable source for specific tRNA gene location, gene sequence, etc. For our final project for BME 160 ‘Biology Research Programming’, we ( Kyle Kaminski and Colin Hortman) worked in collaboration with Todd Lowe to try and improve the GtRNAdb. Currently, there is no presented connections between a tRNA present in the GtRNAdb and research involving that specific tRNA (specific to a particular species). We set out to create a research connection to GtRNAdb tRNA’s under the guidance of Todd Lowe (to the extent of working in his undergraduate research group: the Genome Corps). We have designed a python program that takes in fastA files from the GtRNAdb and searches for relevant research for a specific tRNA of a specific species. Eventually, we hope that there is a clickable